refactor: Replace inline AuthPayload interface with import from dtos in login, refresh-token, and register command handlers for consistency and maintainability
