+++ src/modules/auth/application/dtos/auth-payload.dto.ts
@@ -1,0 +1,7 @@
+import { UserDetailsResult } from 'src/modules/users/application/dtos/user-details.result';
+
+export interface AuthPayload {
+  accessToken: string;
+  refreshToken: string;
+  user: UserDetailsResult;
+}
