+++ src/modules/auth/application/dtos/token-verification-result.dto.ts
@@ -1,0 +1,9 @@
+import { JwtPayload } from '../ports/token.service';
+
+export interface TokenVerificationResult {
+  valid: boolean;
+  payload?: JwtPayload;
+  expired?: boolean;
+  userId?: string;
+  email?: string;
+}
